Simplify this whole thing
Let's not make this too big a deal: It's for rustfix. We should always
error when replacing an already replaced part of the data. Additionally,
we also now track the original data and index into that, while otherwise
using an `Rc<[u8]>` which allows cheaply build up new vectors of parts.